.github/workflows/requirements.txt
# Azure Functions runtime dependencies
azure-functions>=1.11.0
azure-servicebus>=7.11.0

# Core Azure libraries  
azure-core>=1.24.0
azure-identity>=1.12.0

# Development dependencies (uncomment for local development)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
